examples_generator/requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
python-slugify==3.0.3
soupsieve==1.9.3
text-unidecode==1.2
